fix: BufferView: add missing casts in subview
For a fixed capacity Buffer, the `pointer` is picked a small as possible to
hold the required capacity. This means the type of `pointer` varies,
sometimes it matches `size_t`, sometimes it does not...

In the `subview` functions this can cause a narrowing error (`size_t` ->
`pointer`). Given that the pointer size is large enough to hold the
lagest index, we can safely down-cast the arguments (`size_t`) to
`pointer`, as required by the `absolute` helper function.